backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
psycopg2-binary
python-dotenv
chromadb
sentence-transformers
ollama